Change complex*16 to complex(dp)

+++ src/UPSY/validation/unit_tests/ut_mpi_gather_dist_shared_to_all.f90
@@ -617,8 +617,8 @@ contains
     character(len=1024), parameter                :: test_name_local = 'complex_1D'
     character(len=1024)                           :: test_name
     type(type_par_arr_info)                       :: pai
-    complex*16, dimension(:), contiguous, pointer :: d_nih => null()
-    complex*16, dimension(:), contiguous, pointer :: d_tot => null()
+    complex(dp), dimension(:), contiguous, pointer :: d_nih => null()
+    complex(dp), dimension(:), contiguous, pointer :: d_tot => null()
     type(MPI_WIN)                                 :: wd_nih, wd_tot
     logical                                       :: test_result
     integer                                       :: ierr
@@ -677,8 +677,8 @@ contains
     character(len=1024)                             :: test_name
     type(type_par_arr_info)                         :: pai
     integer                                         :: nz
-    complex*16, dimension(:,:), contiguous, pointer :: d_nih => null()
-    complex*16, dimension(:,:), contiguous, pointer :: d_tot => null()
+    complex(dp), dimension(:,:), contiguous, pointer :: d_nih => null()
+    complex(dp), dimension(:,:), contiguous, pointer :: d_tot => null()
     type(MPI_WIN)                                   :: wd_nih, wd_tot
     logical                                         :: test_result
     integer                                         :: ierr
@@ -738,8 +738,8 @@ contains
     character(len=1024)                               :: test_name
     type(type_par_arr_info)                           :: pai
     integer                                           :: nz, nl
-    complex*16, dimension(:,:,:), contiguous, pointer :: d_nih => null()
-    complex*16, dimension(:,:,:), contiguous, pointer :: d_tot => null()
+    complex(dp), dimension(:,:,:), contiguous, pointer :: d_nih => null()
+    complex(dp), dimension(:,:,:), contiguous, pointer :: d_tot => null()
     type(MPI_WIN)                                     :: wd_nih, wd_tot
     logical                                           :: test_result
     integer                                           :: ierr
